add delete team controller

+++ backend/src/services/delete-team.service.ts
@@ -1,7 +1,7 @@
 import { prisma } from '../libs/prisma'
 
-const deleteTeamService = async (teamId: number): Promise<void> => {
-  await prisma.team.delete({ where: { id: teamId } })
+const deleteTeamService = async (userId: number, teamId: number): Promise<void> => {
+  await prisma.team.delete({ where: { id: teamId, userId } })
 }
 
 export { deleteTeamService }
